Restructured components and added navigation
