+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
   mysqldb:
     container_name: tafweb-db
     image: mysql:5.7
-    restart: always
+    restart: unless-stopped
     env_file:
       - ./database/.database.env
     ports:
@@ -18,7 +18,7 @@ services:
     build:
       context: ./backend/
       dockerfile: Dockerfile.backend
-    restart: always
+    restart: unless-stopped
     ports:
       - 8080:8080
     env_file:
@@ -30,7 +30,7 @@ services:
       - files:/files
   app:
     container_name: tafweb-app
-    restart: always
+    restart: unless-stopped
     build:
       context: ./frontend/
       dockerfile: Dockerfile.frontend
