+++ frontend/src/Models/HtmlAudioRef.tsx
@@ -1,5 +1,3 @@
-type HtmlAudioRef = {
+export type HtmlAudioRef = {
     current: HTMLAudioElement | null
 }
-
-export default HtmlAudioRef;
