+++ frontend/src/Models/SongDetail.tsx
@@ -1,8 +1,6 @@
-type SongDetail  = {
+export type SongDetail  = {
     title: string,
     artist: string,
     url: string,
     albumArt: string | null,
 }
-
-export default SongDetail;
