+++ frontend/src/Views/About/About.tsx
@@ -1,0 +1,8 @@
+export default function About()
+{
+    return (
+        <div>
+            <h2>Coming soon</h2>
+        </div>
+    );
+}
