+++ frontend/src/Views/Visuals/Visuals.tsx
@@ -1,0 +1,8 @@
+export default function Visuals()
+{
+    // Empty function show media player can play without any content overlaying it
+    return (
+        <>
+        </>
+    );
+}
